+++ bc/recruitment/templatetags/jobs_search_filters.py
@@ -1,22 +1,32 @@
 from django import template
 
 from bc.recruitment.models import JobCategory
+from bc.recruitment.utils import get_school_and_early_years_count
 
 register = template.Library()
 
 
 @register.inclusion_tag("patterns/molecules/search-filters/search-filters.html")
-def jobs_search_filters(request):
-    job_categories = JobCategory.get_categories_summary().order_by("label")
-    selected_categories = request.GET.getlist("category")
+def jobs_search_filters(request, unfiltered_results):
+    hide_schools_and_early_years = request.GET.get(
+        "hide_schools_and_early_years", False
+    )
+    job_categories = JobCategory.get_categories_summary(unfiltered_results).order_by(
+        "label"
+    )
 
     return {
+        "hide_schools_and_early_years": {
+            "label": "Hide all schools and early years jobs",
+            "count": get_school_and_early_years_count(unfiltered_results),
+            "selected": hide_schools_and_early_years,
+        },
         "filters": [
             {
-                "title": "Show only",
+                "title": "More filtering options",
                 "options": job_categories,
-                "selected": selected_categories,
+                "selected": request.GET.getlist("category"),
                 "key": "category",
-            }
+            },
         ],
     }